Move /ward/visits
